[msg] added message to controller

+++ controller/admin.controller.js
@@ -1,11 +1,15 @@
+// mengimport variabel dari file lain dan dependensi yang dibutuhkan
 const upload = require("./upload-foto-admin").single("foto");
 const adminModel = require("../models/index").admin;
 const Op = require("sequelize").Op;
 const path = require("path");
 const fs = require("fs");
 
+//mengeksport variabel supaya bisa digunakan di file lain
 exports.getAllAdmin = async (req, res) => {
+  // menyiapkan data admin
   let admins = await adminModel.findAll();
+  // mengirim data yang sudah disiapkan 
   return res.json({
     succes: true,
     data: admins,
@@ -13,13 +17,17 @@ exports.getAllAdmin = async (req, res) => {
   });
 };
 
+//mengeksport variabel supaya bisa digunakan di file lain
 exports.findAllAdmin = async (req, res) => {
+  // mengambil kata kunci dari keyword
   let keyword = req.body.keyword;
 
+  // mengambil data admin sesuai dengan kata kunci
   const admins = await adminModel.findAll({
     [Op.or]: [{ name: { [Op.substring]: keyword } }],
   });
 
+  // mengirim data
   return res.json({
     succes: true,
     data: admins,
@@ -27,14 +35,16 @@ exports.findAllAdmin = async (req, res) => {
   });
 };
 
+//mengeksport variabel supaya bisa digunakan di file lain
 exports.addAdmin = async (req, res) => {
+  // mengupload data admin
   upload(req, res, async (err) => {
-    if (err) {
-      return res.json({ message: err });
-    }
-    if (!req.file) {
-      return res.json({ message: "nothing to upload" });
-    }
+    // jika terjadi eror
+    if (err) return res.json({ message: err });
+    // jika tidak ada file yang diapload
+    if (!req.file) return res.json({ message: "nothing to upload" });
+    
+    // menyiapkan data
     let newAdmin = {
       name: req.body.name,
       address: req.body.address,
@@ -43,8 +53,10 @@ exports.addAdmin = async (req, res) => {
       password: req.body.password,
       foto: req.file.filename,
     };
-    adminModel
-      .create(newAdmin)
+
+    // menjalankan proses upload sesuai dengan data
+    adminModel.create(newAdmin)
+    // jika berhasil maka buka promise dengan then
       .then((result) => {
         return res.json({
           succes: true,
@@ -52,6 +64,7 @@ exports.addAdmin = async (req, res) => {
           message: "New Admin has been added",
         });
       })
+    //jika gagal maka buka promise dengan catch
       .catch((error) => {
         return res.json({
           succes: false,
@@ -61,12 +74,15 @@ exports.addAdmin = async (req, res) => {
   });
 };
 
+//mengeksport variabel supaya bisa digunakan di file lain
 exports.updateAdmin = async (req, res) => {
+  // menjalankan proses upload
   upload(req, res, async (error) => {
-    if (error) {
-      return res.json({ message: error });
-    }
+    //jika terjadi error
+    if (error) return res.json({ message: error });
+    //mengambil id admin dengan params
     const idAdmin = req.params.id;
+    // menyiapkan data admin
     const dataAdmin = {
       name: req.body.name,
       address: req.body.address,
@@ -75,33 +91,36 @@ exports.updateAdmin = async (req, res) => {
       password: req.body.password,
       foto: req.file.filename,
     };
+    //jika ada file yang dipilih
     if (req.file) {
+      //menyiapkan data admin untuk tempat menaruh foto 
       const selectedAdmin = await adminModel.findOne({
         where: { id: idAdmin },
       });
+      // menyiapkan data foto
       const oldFoto = selectedAdmin.foto;
-      const pathFotoAdmin = path.join(
-        __dirname,
-        `../foto-admin`,
-        oldFoto
-      );
+      // membuat nama baru untuk file yang disimpan sesuai dengan folder
+      const pathFotoAdmin = path.join(__dirname,`../foto-admin`,oldFoto);
+      // jika sudah ada foto lama maka akan dihapus
       if (fs.existsSync(pathFotoAdmin)) {
         fs.unlink(pathFotoAdmin, (err) => {
           console.log(idAdmin, err);
         });
       }
+      // mengupload foto baru
       dataAdmin.foto = req.file.filename;
     }
-    adminModel
-      .update(dataAdmin, { where: { id: idAdmin } })
+    // menjalankan proses update
+    adminModel.update(dataAdmin, { where: { id: idAdmin } })
+    // jika berhasil maka buka promise dengan then
       .then((field) => {
-        //if succes
         return res.json({
           succes: true,
           data: field,
           message: "Data admin has been update",
         });
       })
+    //jika gagal maka buka promise dengan catch
       .catch((err) => {
         return res.json({
           succes: false,
@@ -111,16 +130,23 @@ exports.updateAdmin = async (req, res) => {
   });
 };
 
+//mengeksport variabel supaya bisa digunakan di file lain
 exports.deleteAdmin = async (req, res) => {
+  // mengambil id dari params
   const idAdmin = req.params.id;
+  // mencari data sesuai dengan params id
   const admin = await adminModel.findOne({ where: { id: idAdmin } });
+  //mengambil foto disimpan divariabel foto lama
   const oldFoto = admin.foto;
+  // menamakan foto sesuai dengan foldernya
   const pathFotoAdmin = path.join(__dirname,`../foto-admin`, oldFoto);
+  // jika foto ada maka akan di hapus
   if (fs.existsSync(pathFotoAdmin)) {
     fs.unlinkSync(pathFotoAdmin, (error) => console.log(error));
   }
-  adminModel
-    .destroy({ where: { id: idAdmin } })
+  // menghapus data admin sesuai dengan id
+  adminModel.destroy({ where: { id: idAdmin } })
+  // jika berhasil maka buka promise dengan then
     .then((fields) => {
       return res.json({
         succes: true,
@@ -128,6 +154,7 @@ exports.deleteAdmin = async (req, res) => {
         message: "Data admin has been deleted",
       });
     })
+  // jika gagal maka akan buka promise menggunakan catch
     .catch((err) => {
       return res.json({
         succes: false,
